backend/requirements.txt
fastapi
uvicorn
google-generativeai
openai
anthropic
